refactor: split upload client internals
